refactor: rename utils to _utils, rename removes_suffix to remove_suffix, and unify snake_case methods
