+++ kloppy/_providers/wyscout.py
@@ -1,6 +1,7 @@
 import json
 from typing import Optional, Union
 
+from kloppy._utils import github_resolve_raw_data_url
 from kloppy.config import get_config
 from kloppy.domain import EventDataset, EventFactory
 from kloppy.infra.serializers.event.wyscout import (
@@ -9,7 +10,6 @@ from kloppy.infra.serializers.event.wyscout import (
     WyscoutInputs,
 )
 from kloppy.io import FileLike, open_as_file
-from kloppy.utils import github_resolve_raw_data_url
 
 
 def load(
