+++ kloppy/_utils/__init__.py
@@ -1,0 +1,9 @@
+from .deprecation import (
+    DeprecatedEnumValue,
+    deprecated,
+)
+from .docstrings import docstring_inherit_attributes
+from .io import Readable, github_resolve_raw_data_url, to_file_object
+from .logging import performance_logging
+from .optional import import_optional_dependency
+from .strings import remove_suffix, to_snake_case
