+++ kloppy/domain/models/code.py
@@ -2,10 +2,10 @@ from dataclasses import dataclass, field
 from datetime import timedelta
 from typing import Union
 
-from kloppy.domain.models.common import DatasetType
-from kloppy.utils import (
+from kloppy._utils import (
     docstring_inherit_attributes,
 )
+from kloppy.domain.models.common import DatasetType
 
 from .common import DataRecord, Dataset
