+++ kloppy/infra/serializers/tracking/signality.py
@@ -5,6 +5,7 @@ import logging
 from typing import IO, NamedTuple, Optional, Union
 import warnings
 
+from kloppy._utils import performance_logging
 from kloppy.domain import (
     AttackingDirection,
     BallState,
@@ -27,7 +28,6 @@ from kloppy.infra.serializers.tracking.deserializer import (
     TrackingDataDeserializer,
 )
 from kloppy.io import FileLike, open_as_file
-from kloppy.utils import performance_logging
 
 logger = logging.getLogger(__name__)
