+++ kloppy/infra/serializers/tracking/skillcorner.py
@@ -4,6 +4,7 @@ import logging
 from typing import IO, NamedTuple, Optional, Union
 import warnings
 
+from kloppy._utils import performance_logging
 from kloppy.domain import (
     AttackingDirection,
     BallState,
@@ -28,7 +29,6 @@ from kloppy.exceptions import DeserializationError
 from kloppy.infra.serializers.tracking.deserializer import (
     TrackingDataDeserializer,
 )
-from kloppy.utils import performance_logging
 
 logger = logging.getLogger(__name__)
