+++ kloppy/infra/serializers/tracking/statsperform.py
@@ -3,6 +3,7 @@ import logging
 from typing import IO, NamedTuple, Optional, Union
 import warnings
 
+from kloppy._utils import performance_logging
 from kloppy.domain import (
     AttackingDirection,
     BallState,
@@ -21,7 +22,6 @@ from kloppy.domain import (
 from kloppy.domain.services.frame_factory import create_frame
 from kloppy.exceptions import DeserializationError
 from kloppy.infra.serializers.event.statsperform.parsers import get_parser
-from kloppy.utils import performance_logging
 
 from .deserializer import TrackingDataDeserializer
